refactor(test): extract changed-scope logic into dedicated module
